higher order component (HOC), reusable logic
