+++ testing/src/Components/CommentBox.js
@@ -1,28 +1,11 @@
 import React, {Component} from 'react'
 import { connect } from 'react-redux'
 import * as actions from 'actions'
-
+import requireAuth from 'Components/requireAuth'
 
 class CommentBox extends Component{
 
-    state = {comment: ''}
-
-    //Component just rendered
-    componentDidMount(){
-
-    }
-
-    componentDidUpdate(){
-        this.shouldNavigateAway();
-    }
-
-    shouldNavigateAway(){
-        if(!this.props.auth){
-            console.log("leave");
-            debugger
-            this.props.history.push('/');
-        }
-    }
+    state = {comment: ''}    
 
     handleChange = (event) =>{
         this.setState({comment: event.target.value})
@@ -53,9 +36,7 @@ class CommentBox extends Component{
     }
 }
 
-function mapStateToProps(state){
-    return {auth: state.auth};
-}
 
-export default connect(mapStateToProps,actions)(CommentBox);
+
+export default connect(null,actions)(requireAuth(CommentBox));
